requirements.txt
flask
flask-cors
requests
msal
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
apscheduler
pandas
click
colorama
pytest
pytest-mock
pytest-cov
cryptography
sqlalchemy
flask-sqlalchemy
flask-migrate
flask-login
email-validator
python-dateutil
beautifulsoup4
lxml
html2text
tqdm
redis
celery
gunicorn
prometheus-client
flask-limiter
tabulate